move keygeneration folder
